requirements.txt
mysql-connector-python
python-dotenv
google-genai
google-search-results
gspread
pandas
sqlalchemy
tabulate
